refactor: 🛠️ Jules05: resolve cross-agent risk and sizing conflicts
- Consolidate institutional risk management logic from `RiskEngine` into `RiskManager`.
- Implement cascading daily loss limits, ATR-based position sizing, and exposure checks in `RiskManager`.
- Standardize structured logging in `MT5Connector` to use keyword arguments exclusively.
- Clean up redundant `risk_engine.py` and update package exports.
- Update integration conflict log and verify system-wide regression tests.
