+++ src/trading/__init__.py
@@ -11,19 +11,21 @@ from src.trading.capital_allocator import (
     RejectionCode,
     StrategyConfig,
 )
+from src.trading.audited_risk_manager import AuditedRiskManager
 from src.trading.mt5_connector import MT5Connector
-from src.trading.risk_engine import RiskDecision, RiskEngine
+from src.trading.risk_manager import RiskDecision, RiskManager
 
 __all__ = [
     "AllocationRequest",
     "AllocationResult",
+    "AuditedRiskManager",
     "BacktestEngine",
     "CapitalAllocator",
     "MT5Connector",
     "PerformanceReport",
     "RejectionCode",
     "RiskDecision",
-    "RiskEngine",
+    "RiskManager",
     "StrategyConfig",
     "TradeSignal",
 ]
