+++ src/trading/mt5_connector.py
@@ -342,7 +342,7 @@ class MT5Connector:
             if isinstance(e, (MT5DataError, MT5ConnectionError)):
                 raise
             # For unexpected exceptions that might be connection-related
-            logger.exception("Unexpected error in get_rates: %s", e)
+            logger.exception("get_rates_unexpected_error", error=str(e))
             self._is_initialized = False
             raise MT5DataError(f"Unexpected data retrieval error: {e}") from e
 
@@ -392,7 +392,7 @@ class MT5Connector:
         except Exception as e:
             if isinstance(e, (MT5DataError, MT5ConnectionError)):
                 raise
-            logger.exception("Unexpected error in get_ticks_range: %s", e)
+            logger.exception("get_ticks_range_unexpected_error", error=str(e))
             self._is_initialized = False
             raise MT5DataError(f"Unexpected tick range retrieval error: {e}") from e
 
@@ -430,7 +430,9 @@ class MT5Connector:
                     err_code, err_desc = mt5.last_error()
                     if err_code in [-1, 10001, 10002, 10003, 10004]:
                         logger.warning(
-                            "MT5 connection failure detected (code %d). Resetting...", err_code
+                            "mt5_connection_failure_detected",
+                            code=err_code,
+                            action="resetting",
                         )
                         self._is_initialized = False
                     raise MT5DataError(f"Failed to copy rates range: {err_desc} (code: {err_code})")
@@ -451,7 +453,7 @@ class MT5Connector:
         except Exception as e:
             if isinstance(e, (MT5DataError, MT5ConnectionError)):
                 raise
-            logger.exception("Unexpected error in get_rates_range: %s", e)
+            logger.exception("get_rates_range_unexpected_error", error=str(e))
             self._is_initialized = False
             raise MT5DataError(f"Unexpected data range retrieval error: {e}") from e
 
@@ -479,7 +481,9 @@ class MT5Connector:
                     err_code, err_desc = mt5.last_error()
                     if err_code in [-1, 10001, 10002, 10003, 10004]:
                         logger.warning(
-                            "MT5 connection failure detected (code %d). Resetting...", err_code
+                            "mt5_connection_failure_detected",
+                            code=err_code,
+                            action="resetting",
                         )
                         self._is_initialized = False
                     is_retriable = err_code not in [-2, -5]
@@ -498,7 +502,7 @@ class MT5Connector:
         except Exception as e:
             if isinstance(e, (MT5DataError, MT5ConnectionError)):
                 raise
-            logger.exception("Unexpected error in get_tick: %s", e)
+            logger.exception("get_tick_unexpected_error", error=str(e))
             self._is_initialized = False
             raise MT5DataError(f"Unexpected tick retrieval error: {e}") from e
