+++ src/trading/risk_manager.py
@@ -20,7 +20,9 @@ from __future__ import annotations
 import logging
 from dataclasses import dataclass, field
 from datetime import date
-from typing import Dict, Optional
+from typing import Any, Dict, List, Optional
+
+import pandas as pd
 
 from src.core.config import TradingConfig
 from src.core.monitor import Monitor
@@ -53,6 +55,15 @@ class DailyStats:
     consecutive_losses: int = 0
 
 
+@dataclass
+class RiskDecision:
+    """Decision details from the RiskManager."""
+
+    is_approved: bool
+    reason: str = ""
+    adjusted_lot_size: float = 0.0
+
+
 class RiskManager:
     """
     Central risk authority.
@@ -81,30 +92,23 @@ class RiskManager:
         signal: TradeSignal,
         signal_id: Optional[int] = None,
         model_health: Optional[dict] = None,
+        open_positions: Optional[List[Dict[str, Any]]] = None,
+        market_data: Optional[pd.DataFrame] = None,
     ) -> bool:
         """
         Run the full 8-layer risk filter cascade.
         Returns True only if ALL layers pass.
         """
-        rejection_reason = ""
-        if not self._check_circuit_breaker():
-            rejection_reason = "Circuit breaker active"
-        elif not self._check_daily_loss():
-            rejection_reason = "Daily loss limit reached"
-        elif not self._check_max_positions():
-            rejection_reason = "Max positions reached"
-        elif not self._check_symbol_allocation(signal.symbol):
-            rejection_reason = f"Symbol {signal.symbol} not in portfolio"
-        elif not self._check_minimum_confidence(signal.confidence):
-            rejection_reason = f"Confidence {signal.confidence:.2f} too low"
-        elif not self._check_risk_reward(signal):
-            rejection_reason = "Risk-Reward ratio too low"
-        elif not self._check_consecutive_losses():
-            rejection_reason = "Max consecutive losses reached"
-        elif not self._check_model_health(model_health):
-            rejection_reason = "Model health metrics below threshold"
+        decision = self.validate_signal(
+            signal,
+            market_data=market_data,
+            open_positions=open_positions or [],
+            model_health=model_health,
+        )
 
-        passed = rejection_reason == ""
+        passed = decision.is_approved
+        rejection_reason = decision.reason
+
         if not passed:
             logger.warning(
                 "Signal REJECTED | %s %s | Reason: %s",
@@ -120,6 +124,69 @@ class RiskManager:
                     signal_id=signal_id,
                 )
         return passed
+
+    def validate_signal(
+        self,
+        signal: TradeSignal,
+        market_data: Optional[pd.DataFrame] = None,
+        open_positions: List[Dict[str, Any]] = [],
+        model_health: Optional[dict] = None,
+    ) -> RiskDecision:
+        """
+        Validate a trade signal against the 8-layer cascade.
+        """
+        # Layer 1: Circuit Breakers (Equity Drawdown)
+        if not self._check_circuit_breaker():
+            return RiskDecision(False, "Hard drawdown limit reached")
+
+        # Layer 2: Daily Loss Limits (Institutional Cascading)
+        if self.get_daily_loss_level() >= 4:
+            return RiskDecision(False, "Daily loss limit reached (Level 4)")
+
+        # Layer 3: Activity Limits
+        if not self._check_max_positions():
+            return RiskDecision(False, "Max concurrent positions reached")
+        if not self._check_consecutive_losses():
+            return RiskDecision(False, "Max consecutive losses reached")
+
+        # Layer 4: Exposure Limits
+        if open_positions:
+            if not self._check_directional_exposure(signal, open_positions):
+                return RiskDecision(False, "Max directional exposure reached (30%)")
+            if market_data is not None and not self._check_total_notional(
+                signal, open_positions, market_data
+            ):
+                return RiskDecision(False, "Total notional exposure exceeds equity")
+
+        # Layer 5: Symbol Allocation
+        if not self._check_symbol_allocation(signal.symbol):
+            return RiskDecision(False, f"Symbol {signal.symbol} not in approved list")
+
+        # Layer 6: Prediction Limits
+        if not self._check_minimum_confidence(signal.confidence):
+            return RiskDecision(
+                False, f"Confidence {signal.confidence:.2f} below {self.cfg.min_confidence}"
+            )
+
+        # Layer 7: Risk-Reward Validation (Min 1.5 R:R)
+        if not self._check_risk_reward(signal):
+            return RiskDecision(False, "Risk-Reward ratio below 1.5")
+
+        # Layer 8: Model Health
+        if not self._check_model_health(model_health):
+            return RiskDecision(False, "Model health metrics below threshold")
+
+        # Calculate lot size (Defaults to ATR-based if market_data present)
+        if market_data is not None and not market_data.empty:
+            adjusted_lots = self.calculate_position_size_atr(signal.symbol, market_data)
+        else:
+            # Fallback to config-based or Kelly if needed
+            adjusted_lots = signal.lot_size
+
+        if adjusted_lots < self.cfg.min_lot_size:
+            return RiskDecision(False, f"Calculated lot size {adjusted_lots} below minimum")
+
+        return RiskDecision(True, "Approved", adjusted_lots)
 
     def size_position(
         self,
@@ -147,6 +214,47 @@ class RiskManager:
             lot_size,
         )
         return lot_size
+
+    def calculate_position_size_atr(self, symbol: str, market_data: pd.DataFrame) -> float:
+        """
+        ATR-based position sizing according to RISK_LIMITS.md.
+        """
+        if market_data.empty or "atr" not in market_data.columns:
+            return self.cfg.min_lot_size
+
+        current_atr = market_data["atr"].iloc[-1]
+        avg_atr = market_data["atr"].tail(8640).mean()  # Approx 30 days of M5
+
+        vol_multiplier = 1.0
+        ratio = current_atr / avg_atr if avg_atr > 0 else 1.0
+
+        if ratio > self.cfg.volatility_extreme_threshold:
+            return 0.0
+        elif ratio > self.cfg.volatility_very_high_threshold:
+            vol_multiplier = 0.5
+        elif ratio > self.cfg.volatility_high_threshold:
+            vol_multiplier = 0.75
+
+        loss_multiplier = self.get_size_multiplier_from_loss()
+        total_multiplier = vol_multiplier * loss_multiplier
+
+        if total_multiplier <= 0:
+            return 0.0
+
+        # Sizing: risk 1% (cfg.risk_per_trade) of balance
+        risk_amount = self.balance * self.cfg.risk_per_trade
+        # ATR * 100 converts gold ATR to $ per lot
+        lot_size = (risk_amount / (current_atr * 100)) * total_multiplier
+
+        # Cap at Max Position Size (10% of equity)
+        max_notional = self.balance * self.cfg.max_position_size_pct
+        price = market_data["close"].iloc[-1]
+        max_lots = max_notional / (price * 100)
+
+        final_lots = min(lot_size, max_lots)
+        final_lots = max(self.cfg.min_lot_size, round(final_lots, 2))
+
+        return final_lots
 
     def update_equity(self, current_equity: float) -> None:
         """Call after every closed trade or on heartbeat."""
@@ -229,13 +337,62 @@ class RiskManager:
         return True
 
     def _check_daily_loss(self) -> bool:
-        if self.daily.peak_equity == 0:
-            return True
+        """Legacy check maintained for compatibility."""
+        return self.get_daily_loss_level() < 4
+
+    def get_daily_loss_level(self) -> int:
+        """
+        Institutional Daily Loss Level (0-4).
+        """
+        if self.daily.peak_equity <= 0 or self.daily.realised_pnl >= 0:
+            return 0
+
         loss_pct = abs(self.daily.realised_pnl) / self.daily.peak_equity
-        if self.daily.realised_pnl < 0 and loss_pct >= self.cfg.max_daily_loss:
-            logger.warning("Daily loss limit hit: %.1f%%", loss_pct * 100)
-            return False
-        return True
+
+        if loss_pct >= self.cfg.max_daily_loss:
+            return 4
+        if loss_pct >= (getattr(self.cfg, "daily_loss_lvl3", self.cfg.max_daily_loss * 0.75)):
+            return 3
+        if loss_pct >= (getattr(self.cfg, "daily_loss_lvl2", self.cfg.max_daily_loss * 0.5)):
+            return 2
+        if loss_pct >= (getattr(self.cfg, "daily_loss_lvl1", self.cfg.max_daily_loss * 0.25)):
+            return 1
+        return 0
+
+    def get_size_multiplier_from_loss(self) -> float:
+        """Multiplier based on daily loss level."""
+        level = self.get_daily_loss_level()
+        mapping = {0: 1.0, 1: 1.0, 2: 0.5, 3: 0.25, 4: 0.0}
+        return mapping.get(level, 0.0)
+
+    def _check_directional_exposure(
+        self, signal: TradeSignal, open_positions: List[Dict[str, Any]]
+    ) -> bool:
+        """Layer 4: 30% net directional exposure."""
+        net_lots = 0.0
+        for pos in open_positions:
+            vol = pos.get("volume", 0.0)
+            if pos.get("type") == 0:  # BUY
+                net_lots += vol
+            else:  # SELL
+                net_lots -= vol
+
+        net_lots += self.cfg.min_lot_size if signal.direction > 0 else -self.cfg.min_lot_size
+        price_estimate = 2300.0  # Gold estimate
+        notional = abs(net_lots) * price_estimate * 100
+        exposure_pct = notional / self.balance if self.balance > 0 else 1.0
+
+        return exposure_pct <= self.cfg.max_single_direction_pct
+
+    def _check_total_notional(
+        self, signal: TradeSignal, open_positions: List[Dict[str, Any]], market_data: pd.DataFrame
+    ) -> bool:
+        """Layer 4: Total notional < 100% equity (configured)."""
+        total_lots = sum(pos.get("volume", 0.0) for pos in open_positions) + self.cfg.min_lot_size
+        price = market_data["close"].iloc[-1] if not market_data.empty else 2300.0
+        total_notional = total_lots * price * 100
+        limit_pct = getattr(self.cfg, "max_total_notional_pct", 1.0)
+        return total_notional < (self.balance * limit_pct)
 
     def _check_max_positions(self) -> bool:
         if len(self.open_positions) >= self.cfg.max_positions:
@@ -268,4 +425,4 @@ class RiskManager:
         return True
 
 
-__all__ = ["ALLOCATION_WEIGHTS", "DailyStats", "RiskManager"]
+__all__ = ["ALLOCATION_WEIGHTS", "DailyStats", "RiskDecision", "RiskManager"]
