+++ tests/test_institutional_integration.py
@@ -90,6 +90,8 @@ def test_capital_and_risk_integration(trade_logger):
         cfg.max_daily_loss = 0.05
         cfg.max_positions = 3
         cfg.max_losing_streak = 5
+        cfg.min_confidence = 0.55
+        cfg.min_lot_size = 0.01
         mock_get_cfg.return_value = cfg
 
         # 1. Capital Allocation (Jules04)
