+++ tests/test_risk_engine_new.py
@@ -1,19 +1,19 @@
 import pytest
 import pandas as pd
-from src.trading.risk_engine import RiskEngine
+from src.trading.risk_manager import RiskManager
 from src.core.config import TradingConfig
 
 @pytest.fixture
 def risk_engine():
     cfg = TradingConfig(MT5_PASSWORD="test", MT5_SERVER="test")
-    return RiskEngine(cfg, 10000.0)
+    return RiskManager(cfg, 10000.0)
 
 def test_drawdown_breaker(risk_engine):
-    risk_engine.update_metrics(6000.0) # 40% drawdown
-    assert not risk_engine._check_drawdown_breaker()
+    risk_engine.update_equity(6000.0) # 40% drawdown
+    assert not risk_engine._check_circuit_breaker()
 
 def test_daily_loss_breaker(risk_engine):
-    risk_engine.update_metrics(10000.0, realized_pnl=-600.0) # 6% loss
+    risk_engine.record_pnl(-600.0) # 6% loss
     assert risk_engine.get_daily_loss_level() >= 4
 
 def test_calculate_position_size(risk_engine):
@@ -21,7 +21,7 @@ def test_calculate_position_size(risk_engine):
         "atr": [1.0] * 100,
         "close": [2300.0] * 100
     })
-    size = risk_engine.calculate_position_size("XAUUSD", data)
+    size = risk_engine.calculate_position_size_atr("XAUUSD", data)
     assert size >= 0.01
 
 def test_validate_signal_rejection(risk_engine):
